rename rootSlice to userSlice, remove obsolete currentRoot from redux, update imports
